[EC-1014] Changed Organization.Status from SMALLINT to TINYINT

+++ src/Sql/dbo/Stored Procedures/Organization_Update.sql
@@ -44,7 +44,7 @@
     @UseScim BIT = 0,
     @UseCustomPermissions BIT = 0,
     @UseSecretsManager BIT = 0,
-    @Status SMALLINT = 0
+    @Status TINYINT = 0
 AS
 BEGIN
     SET NOCOUNT ON
